refactor to add secrets module
