+++ handlers/zuora-salesforce-link-remover/src/getBillingAccounts.ts
@@ -1,7 +1,5 @@
-import {
-	GetSecretValueCommand,
-	SecretsManagerClient,
-} from '@aws-sdk/client-secrets-manager';
+import type { ApiUserSecret, ConnectedAppSecret } from "./secrets";
+import { getSecretValue } from "./secrets";
 
 export async function handler() {
 	const connectedAppSecretName = 'DEV/Salesforce/ConnectedApp/AwsConnectorSandbox'; //TODO in future pr: add a function that returns appropriate secret depending on env
@@ -16,37 +14,3 @@ export async function handler() {
 	
 	return 'abcdef';
 }
-
-export async function getSecretValue<T>(secretName: string): Promise<T>{
-	try {
-		const secretsManagerClient = new SecretsManagerClient({
-			region: process.env.region,
-		});
-
-		const command = new GetSecretValueCommand({
-			SecretId: secretName,
-		});
-
-		const response = await secretsManagerClient.send(command);
-		
-		if (!response.SecretString) {
-			throw new Error(`No secret found with name:$secretName`);
-		}
-
-		return JSON.parse(response.SecretString) as T;
-	} catch (error) {
-		console.error('error:', error);
-		throw error;
-	}
-}
-
-type ConnectedAppSecret = {
-	name: string;
-	authUrl: string;
-};
-
-type ApiUserSecret = {
-	username: string;
-	password: string;
-	token: string;
-};
